+++ activerecord/test/cases/timestamp_test.rb
@@ -92,6 +92,31 @@ class TimestampTest < ActiveRecord::TestCase
     assert_nothing_raised { Car.first.touch }
   end
 
+  def test_touching_a_no_touching_object
+    Developer.no_touching do
+      @developer.touch
+    end
+    assert_equal @previously_updated_at, @developer.updated_at
+  end
+
+  def test_touching_related_objects
+    @owner = Owner.first
+    @previously_updated_at = @owner.updated_at
+
+    Owner.no_touching do
+      @owner.pets.first.touch
+    end
+
+    assert_equal @previously_updated_at, @owner.updated_at
+  end
+
+  def test_global_no_touching
+    ActiveRecord::Base.no_touching do
+      @developer.touch
+    end
+    assert_equal @previously_updated_at, @developer.updated_at
+  end
+
   def test_saving_a_record_with_a_belongs_to_that_specifies_touching_the_parent_should_update_the_parent_updated_at
     pet   = Pet.first
     owner = pet.owner